requirements.txt
streamlit~=1.18.1
pandas~=1.5.3
requests~=2.28.2
Pillow~=9.4.0
plotly~=5.13.0
networkx~=3.0
dash~=2.8.1
dash_cytoscape~=0.3.0
matplotlib~=3.7.0
pyvis~=0.3.1
ipython~=8.10.0